medication-rag/requirements.txt
# Medication RAG System Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LlamaIndex (RAG)
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai>=0.1.0

# OpenAI
openai>=1.0.0
